Move WidgetLocation out of element_extensions.dart
- Move WidgetLocation, WidgetLocationFile and the Element extension that
  resolves them into widget_location.dart. element_extensions.dart is now
  only about traversing the Element tree.
- Close WidgetLocation.toString(), it was missing its brace.
- Document every public member in act_tap.dart and drop the file-wide
  public_member_api_docs ignore, so the next exported type cannot ship
  undocumented in silence.
- Import dartx instead of dartx_io, nothing in the file uses dart:io.
- Share one definition of "outside the viewport" between the throwing
  validateViewBounds and the reporting _inspectViewBounds. Both did the
  same cheap arithmetic and built the same message.
- Drop validateViewBounds' unused throwWhenNotInViewport parameter and its
  ignored return value.
- Delete _selfAndParents, Element.parents already starts at the element.
  Comment the two call sites that rely on it.
- Hand out unmodifiable lists from the exported report types.
- Document TapInspection as a snapshot that goes stale after a pump, and
  fix the dangling [target] reference on TapZeroSizeReason.
